Do not call mount.RecursiveUnmount() on Windows
