+++ daemon/daemon_unix.go
@@ -1708,3 +1708,7 @@ func (daemon *Daemon) RawSysInfo(quiet bool) *sysinfo.SysInfo {
 	}
 	return sysinfo.New(quiet, opts...)
 }
+
+func recursiveUnmount(target string) error {
+	return mount.RecursiveUnmount(target)
+}
