+++ testutil/daemon/daemon.go
@@ -24,7 +24,6 @@ import (
 	"github.com/docker/docker/testutil/request"
 	"github.com/docker/go-connections/sockets"
 	"github.com/docker/go-connections/tlsconfig"
-	"github.com/moby/sys/mount"
 	"github.com/pkg/errors"
 	"gotest.tools/v3/assert"
 )
@@ -808,15 +807,6 @@ func (d *Daemon) Info(t testing.TB) types.Info {
 	assert.NilError(t, err)
 	assert.NilError(t, c.Close())
 	return info
-}
-
-// cleanupMount unmounts the daemon root directory, or logs a message if
-// unmounting failed.
-func cleanupMount(t testing.TB, d *Daemon) {
-	t.Helper()
-	if err := mount.Unmount(d.Root); err != nil {
-		d.log.Logf("[%s] unable to unmount daemon root (%s): %v", d.id, d.Root, err)
-	}
 }
 
 // cleanupRaftDir removes swarmkit wal files if present
